requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==3.1.6
django-bootstrap4==2.3.1
djangorestframework==3.12.2
Pillow==8.1.0
whitenoise==5.2.0
